Preventing retentionTime and retentionCount to be set to 0 (#2790)
* Preventing retentionTime and retentionCount to be set to 0 for dataset

* Filter out the null(default) value during serialization

* Address comments

---------

+++ ambry-frontend/src/test/java/com/github/ambry/frontend/NamedBlobPutHandlerTest.java
@@ -110,10 +110,11 @@ public class NamedBlobPutHandlerTest {
       REF_CONTAINER = REF_ACCOUNT.getContainerById(Container.DEFAULT_PRIVATE_CONTAINER_ID);
       REF_CONTAINER_WITH_TTL_REQUIRED = REF_ACCOUNT.getContainerById(Container.DEFAULT_PUBLIC_CONTAINER_ID);
       ACCOUNT_SERVICE.addDataset(
-          new DatasetBuilder(REF_ACCOUNT.getName(), REF_CONTAINER.getName(), DATASET_NAME).setRetentionTimeInSeconds(-1)
+          new DatasetBuilder(REF_ACCOUNT.getName(), REF_CONTAINER.getName(), DATASET_NAME).setRetentionTimeInSeconds(
+                  (long) -1)
               .build());
       ACCOUNT_SERVICE.addDataset(new DatasetBuilder(REF_ACCOUNT.getName(), REF_CONTAINER_WITH_TTL_REQUIRED.getName(),
-          DATASET_NAME).setRetentionTimeInSeconds(-1).build());
+          DATASET_NAME).setRetentionTimeInSeconds((long) -1).build());
       ACCOUNT_SERVICE.addDatasetVersion(REF_ACCOUNT.getName(), REF_CONTAINER.getName(), DATASET_NAME, VERSION, -1,
           System.currentTimeMillis(), false, null);
       ACCOUNT_SERVICE.addDatasetVersion(REF_ACCOUNT.getName(), REF_CONTAINER_WITH_TTL_REQUIRED.getName(), DATASET_NAME,
